pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.0.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=1.0.5",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pypdf>=3.0.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
]
